Bug fixes and Cleaner syntax
