+++ bonobo/example/compiler.dart
@@ -1,0 +1,7 @@
+import 'package:bonobo/bonobo.dart';
+import 'package:file/local.dart';
+
+main(List<String> args) async {
+  var analyzer = await analyze(new CompileOptions());
+  print(analyzer.module.scope.allVariables.map((v) => v.name));
+}
