+++ bonobo/example/parser.dart
@@ -1,0 +1,12 @@
+import 'package:bonobo/bonobo.dart';
+import 'package:file/local.dart';
+
+main(List<String> args) async {
+  const options = const CompileOptions();
+  CompilerInput input = await CompilerInput.fromOptions(options);
+  var scanner = new Scanner(input.contents, sourceUrl: input.uri)..scan();
+  print(scanner.errors);
+  print(scanner.tokens.join('\n\n'));
+  var parser = new Parser(scanner);
+  CompilationUnitContext compilationUnit = parser.parseCompilationUnit();
+}
