+++ bonobo/example/scanner.dart
@@ -1,0 +1,9 @@
+import 'package:bonobo/bonobo.dart';
+
+main(List<String> args) async {
+  const options = const CompileOptions();
+  CompilerInput input = await CompilerInput.fromOptions(options);
+  var scanner = new Scanner(input.contents, sourceUrl: input.uri)..scan();
+  print(scanner.errors);
+  print(scanner.tokens.join('\n\n'));
+}
