+++ bonobo/lib/analysis/num.dart
@@ -14,8 +14,7 @@ class _BonoboNumType extends BonoboInheritedType {
   @override
   BonoboType binaryOp(
       Token operator, BonoboType other, BonoboAnalyzer analyzer) {
-        if (other.isAssignableTo(this))
-        return this;
+    if (other.isAssignableTo(this)) return this;
     return unsupportedBinaryOperator(operator, other, analyzer);
   }
 }
