+++ bonobo/lib/analysis/object.dart
