+++ bonobo/lib/analysis/scope.dart
@@ -1,12 +1,14 @@
 part of bonobo.src.analysis;
 
-Future<SymbolTable<BonoboObject>> createRootScope(AnalysisContext context) async {
+Future<SymbolTable<BonoboObject>> createRootScope(
+    AnalysisContext context) async {
   var scope = new SymbolTable<BonoboObject>();
 
   return scope;
 }
 
-Future<Map<String, BonoboType>> createGlobalTypes(AnalysisContext context) async {
+Future<Map<String, BonoboType>> createGlobalTypes(
+    AnalysisContext context) async {
   return {
     BonoboType.Function$.name: BonoboType.Function$,
     BonoboType.Num.name: BonoboType.Num,
