+++ bonobo/lib/analysis/string.dart
