+++ bonobo/lib/ast/compilation_unit.dart
@@ -3,5 +3,5 @@ part of bonobo.src.ast;
 class CompilationUnitContext extends AstNode {
   final List<FunctionContext> functions;
 
-  CompilationUnitContext(FileSpan span, this.functions):super(span, []);
+  CompilationUnitContext(FileSpan span, this.functions) : super(span, []);
 }
