--- bonobo/lib/commands/compiler.dart
@@ -1,1 +1,0 @@
-part of bonobo.src.commands;
