+++ bonobo/lib/frontend/parser/parser.dart
@@ -152,11 +152,12 @@ class Parser extends _Parser {
       while (modifier != null) {
         modifiers.add(modifier.type);
         span = span == null ? modifier.span : span.expand(modifier.span);
+        modifier = parseModifier();
       }
     }
 
     comments ??= parseComments();
-    f ??= nextToken(TokenType.f);
+    f ??= nextToken(TokenType.func);
     if (f == null) return null;
 
     span = span == null ? f.span : span.expand(f.span);
@@ -532,42 +533,42 @@ class InfixParselet {
 class BinaryParselet extends InfixParselet {
   BinaryParselet(int precedence)
       : super(precedence,
-          (parser, left, token, comments, inVariableDeclaration) {
-        var span = left.span.expand(token.span), lastSpan = span;
-        var equals = token.type == TokenType.equals
-            ? null
-            : parser.nextToken(TokenType.equals)?.span;
+            (parser, left, token, comments, inVariableDeclaration) {
+          var span = left.span.expand(token.span), lastSpan = span;
+          var equals = token.type == TokenType.equals
+              ? null
+              : parser.nextToken(TokenType.equals)?.span;
 
-        if (equals != null) {
-          span = span.expand(lastSpan = equals);
-        }
+          if (equals != null) {
+            span = span.expand(lastSpan = equals);
+          }
 
-        var right = parser.parseExpression(0, inVariableDeclaration);
+          var right = parser.parseExpression(0, inVariableDeclaration);
 
-        if (right == null) {
-          parser.errors.add(new BonoboError(BonoboErrorSeverity.error,
-              "Missing expression after '${lastSpan.text}'.", lastSpan));
-          return null;
-        }
+          if (right == null) {
+            parser.errors.add(new BonoboError(BonoboErrorSeverity.error,
+                "Missing expression after '${lastSpan.text}'.", lastSpan));
+            return null;
+          }
 
-        span = span.expand(right.span);
+          span = span.expand(right.span);
 
-        if (equals != null || token.type == TokenType.equals) {
-          return new AssignmentExpressionContext(
+          if (equals != null || token.type == TokenType.equals) {
+            return new AssignmentExpressionContext(
+              left,
+              token,
+              right,
+              span,
+              []..addAll(comments)..addAll(right.comments),
+            );
+          }
+
+          return new BinaryExpressionContext(
             left,
             token,
             right,
             span,
             []..addAll(comments)..addAll(right.comments),
           );
-        }
-
-        return new BinaryExpressionContext(
-          left,
-          token,
-          right,
-          span,
-          []..addAll(comments)..addAll(right.comments),
-        );
-      });
+        });
 }
