+++ example/bad_style/main.bnb
@@ -1,6 +1,6 @@
 // Bonobo is a very opinionated language, and the formatter is also very opinionated.
 //
 // Take a look... Run `bonobo format bad_style.bnb` and see the output!
-f preferSingleQuotes => "These will be replaced with single quotes!!!"
+func preferSingleQuotes => "These will be replaced with single quotes!!!"
 
-f occassionalDoubleQuotes => "Single quotes are not inserted if the string already contains them. '''"
+func occassionalDoubleQuotes => "Single quotes are not inserted if the string already contains them. '''"
