+++ example/modules/main.bnb
@@ -1,1 +1,1 @@
-f main => print foo.bar()
+func main => print foo.bar()
