+++ example/type_inference/main.bnb
@@ -4,12 +4,12 @@
 //
 // In practice, you will usually not have to even provide
 // type annotations.
-f main => 0
+func main => 0
 
 // Bonobo automatically recognizes the return type to be `Num`.
-f someNum(a: Num, b: Num) => 24.0
+func someNum(a: Num, b: Num) => 24.0
 
 // Bonobo can also infer the types of parameters.
 //
 // `a` is recognized to be a `Num`.
-f someOtherNum(a) => someNum(a, 35)
+func someOtherNum(a) => someNum(a, 35)
